+++ src/armature.h
@@ -32,7 +32,7 @@ private:
 	class Slot_GD *slot{ nullptr };
 	Ref<Texture2D> texture_override;
 
-	class DragonBonesArmatureDisplay *armature{ nullptr };
+	class DragonBonesArmatureDisplay *armature_display{ nullptr };
 	friend class DragonBonesFactory;
 
 protected:
@@ -89,8 +89,8 @@ public:
 			}
 			auto display = slot->getDisplayList()[slot->getDisplayIndex()];
 			if (display.second == dragonBones::DisplayType::Armature) {
-				dragonBones::Armature *armature = static_cast<dragonBones::Armature *>(display.first);
-				DragonBonesArmature *convertedDisplay = static_cast<DragonBonesArmature *>(armature->getDisplay());
+				dragonBones::Armature *armature_display = static_cast<dragonBones::Armature *>(display.first);
+				DragonBonesArmature *convertedDisplay = static_cast<DragonBonesArmature *>(armature_display->getDisplay());
 				if constexpr (std::is_invocable_r_v<bool, Func, DragonBonesArmature *>) {
 					if (p_action(convertedDisplay)) {
 						break;
@@ -120,7 +120,7 @@ public:
 	virtual void append_draw_data(VMap<int, LocalVector<DrawData>> &r_data, const Transform2D &p_base_transfrom = Transform2D()) const override;
 
 public:
-	bool is_valid() const { return armature_instance && armature; }
+	bool is_valid() const { return armature_instance && armature_display; }
 
 	/* METHOD BINDINGS */
 	static void _bind_methods();
